Address PR review feedback (#7977)
- reset no-progress evidence when rebasing to a new run

- pin structured-family replay identity and changing-output coverage

- reuse launched call signatures without recording unlaunched calls

- split stop strategy tests and consolidate repeated fixtures
