+++ crates/loop/ironclaw_agent_loop/src/state.rs
@@ -419,7 +419,9 @@ impl LoopExecutionState {
     /// transcript/result refs in the payload are also owned by the source run;
     /// carrying them into the retry would make the retry's terminal exit claim
     /// foreign-run evidence. Reset these run-owned fields and let the retry host
-    /// produce its own refs.
+    /// produce its own refs. The output-digest observations and terminal
+    /// warning/control state are likewise run-owned and are reset below so the
+    /// retry cannot inherit a no-progress strike.
     ///
     /// Gate-bound resume state (`last_gate`, `pending_approval_resume`,
     /// `pending_auth_resume`) is deliberately NOT cleared here: this same path
@@ -449,6 +451,12 @@ impl LoopExecutionState {
         // does any work. Reset it here so the retry starts its own budget
         // window. (Same-run gate resumes return early above, preserving it.)
         self.budget_ledger = BudgetLedger::fresh_for_run();
+        // No-progress observations and terminal warnings are also scoped to
+        // the source run. Carrying either across a retry would let the new
+        // run inherit a prior strike and terminate without earning it.
+        self.seen_capability_output_digests = BoundedRing::new();
+        self.terminal_warning_state = TerminalWarningState::default();
+        self.stop_state.trailing_no_progress_results = 0;
         self
     }
 }
@@ -1092,6 +1100,51 @@ mod tests {
     }
 
     #[test]
+    fn rebase_for_run_drops_no_progress_observations_and_warning_for_a_new_run() {
+        let source_context = test_run_context();
+        let target_context = test_run_context();
+        let mut state = LoopExecutionState::initial_for_run(&source_context);
+        state
+            .seen_capability_output_digests
+            .push(CapabilityOutputObservation {
+                signature: CapabilityCallSignature::from_call(
+                    CapabilityId::new("demo.echo").expect("valid capability id"),
+                    &json!({"value": 1}),
+                )
+                .expect("valid call signature"),
+                output_digest: ironclaw_loop_contracts::ContentDigest(42),
+            });
+        assert!(
+            state
+                .terminal_warning_state
+                .schedule(TerminalWarningObservation::no_progress(Some(8), None))
+        );
+        state.stop_state.trailing_no_progress_results = 1;
+        state.pending_model_error_observation = Some(ModelErrorRecoveryObservation::transient());
+        state.pending_model_retry_directive = Some(PendingModelRetryDirective::RepairInvalidOutput);
+        state.recovery_state =
+            RecoveryStrategyState::with_attempts_for(RecoveryAttemptClass::ModelTransient, 2);
+
+        let rebased = state.clone().rebase_for_run(&target_context);
+
+        assert!(rebased.seen_capability_output_digests.is_empty());
+        assert_eq!(
+            rebased.terminal_warning_state,
+            TerminalWarningState::default()
+        );
+        assert_eq!(rebased.stop_state.trailing_no_progress_results, 0);
+        assert_eq!(
+            rebased.pending_model_error_observation,
+            state.pending_model_error_observation
+        );
+        assert_eq!(
+            rebased.pending_model_retry_directive,
+            state.pending_model_retry_directive
+        );
+        assert_eq!(rebased.recovery_state, state.recovery_state);
+    }
+
+    #[test]
     fn rebase_for_run_resets_per_run_budget_counters_for_a_different_run() {
         // A retry rebases the failed run's checkpoint onto a fresh TurnRunId.
         // run_started_at/model_calls_made/capability_invocations_made are
@@ -1150,6 +1203,22 @@ mod tests {
         state
             .budget_ledger
             .set_capability_invocations_made_for_test(7);
+        state
+            .seen_capability_output_digests
+            .push(CapabilityOutputObservation {
+                signature: CapabilityCallSignature::from_call(
+                    CapabilityId::new("demo.echo").expect("valid capability id"),
+                    &json!({"value": 1}),
+                )
+                .expect("valid call signature"),
+                output_digest: ironclaw_loop_contracts::ContentDigest(42),
+            });
+        assert!(
+            state
+                .terminal_warning_state
+                .schedule(TerminalWarningObservation::no_progress(Some(8), None))
+        );
+        state.stop_state.trailing_no_progress_results = 1;
 
         let rebased = state.clone().rebase_for_run(&context);
